+++ packages/s2-core/__tests__/unit/facet/pivot-facet-spec.ts
@@ -132,6 +132,7 @@ jest.mock('@/data-set/pivot-data-set', () => {
         getFieldsAndPivotMetaByField:
           actualPivotDataSet.prototype.getFieldsAndPivotMetaByField,
         displayFormattedValueMap: new Map(),
+        dimensionValuesCache: new Map(),
       };
     }),
   };