+++ packages/s2-core/src/cell/corner-cell.ts
@@ -1,5 +1,5 @@
 import type { PointLike } from '@antv/g';
-import { last } from 'lodash';
+import { last, values } from 'lodash';
 import {
   CellType,
   KEY_GROUP_CORNER_RESIZE_AREA,
@@ -61,10 +61,16 @@ export class CornerCell extends HeaderCell<CornerHeaderConfig> {
       return;
     }
 
-    const { collapseAll } = this.spreadsheet.options.style?.rowCell!;
+    const { collapseFields = {} } = this.spreadsheet.options.style?.rowCell!;
     const { size = 0 } = this.getStyle()!.icon!;
     const { fill } = this.getTextStyle();
     const area = this.getBBoxByType(CellClipBox.CONTENT_BOX);
+    const rootRowNodes = this.spreadsheet.facet.getRowNodes(0);
+    // 任意一级节点的 icon 展开/收起, 同步更新角头 icon 的状态
+    const isAllCollapsed =
+      values(collapseFields!).filter(Boolean).length === rootRowNodes.length ||
+      rootRowNodes.every((node) => node.isCollapsed);
+    const isCollapsed = isAllCollapsed;
 
     this.treeIcon = renderTreeIcon({
       group: this,
@@ -75,12 +81,12 @@ export class CornerCell extends HeaderCell<CornerHeaderConfig> {
         height: size,
         fill,
       },
-      isCollapsed: collapseAll,
+      isCollapsed,
       onClick: () => {
         this.spreadsheet.facet.resetScrollY();
         this.spreadsheet.emit(
           S2Event.ROW_CELL_ALL_COLLAPSED__PRIVATE,
-          collapseAll!,
+          isCollapsed,
         );
       },
     });