+++ packages/s2-core/src/facet/pivot-facet.ts
@@ -16,7 +16,7 @@ import {
   size,
   sumBy,
 } from 'lodash';
-import { ColCell, RowCell, SeriesNumberCell } from '../cell';
+import { RowCell, SeriesNumberCell } from '../cell';
 import {
   DEFAULT_TREE_ROW_CELL_WIDTH,
   FRONT_GROUND_GROUP_FROZEN_Z_INDEX,
@@ -36,6 +36,7 @@ import type { PivotDataSet } from '../data-set/pivot-data-set';
 import { renderLine, safeJsonParse } from '../utils';
 import { getDataCellId } from '../utils/cell/data-cell';
 import { getActionIconConfig } from '../utils/cell/header-cell';
+import { getHeaderTotalStatus } from '../utils/dataset/pivot-data-set';
 import { getIndexRangeWithOffsets } from '../utils/facet';
 import { getRowsForGrid } from '../utils/grid';
 import { floor } from '../utils/math';
@@ -112,6 +113,7 @@ export class PivotFacet extends FrozenFacet {
     const isTotals =
       row.isTotals || row.isTotalMeasure || col.isTotals || col.isTotalMeasure;
 
+    const totalStatus = getHeaderTotalStatus(row, col);
     const hideMeasure = options.style?.colCell?.hideValue ?? false;
 
     /*
@@ -129,6 +131,7 @@ export class PivotFacet extends FrozenFacet {
       query: dataQuery,
       rowNode: row,
       isTotals,
+      totalStatus,
     });
 
     const valueField = dataQuery[EXTRA_FIELD]!;
@@ -393,26 +396,16 @@ export class PivotFacet extends FrozenFacet {
   }
 
   private getRowNodeHeight(rowNode: Node): number {
+    if (!rowNode) {
+      return 0;
+    }
+
     const rowCell = new RowCell(rowNode, this.spreadsheet, {
       shallowRender: true,
     });
     const defaultHeight = this.getRowCellHeight(rowNode);
 
     return this.getCellAdaptiveHeight(rowCell, defaultHeight);
-  }
-
-  protected getColNodeHeight(colNode: Node, colsHierarchy: Hierarchy): number {
-    if (!colNode) {
-      return 0;
-    }
-
-    const colCell = new ColCell(colNode, this.spreadsheet, {
-      shallowRender: true,
-    });
-
-    const defaultHeight = this.getDefaultColNodeHeight(colNode, colsHierarchy);
-
-    return this.getCellAdaptiveHeight(colCell, defaultHeight);
   }
 
   /**
@@ -883,8 +876,10 @@ export class PivotFacet extends FrozenFacet {
               valueData,
             ) ?? valueData;
           const cellLabel = `${formattedValue}`;
-          const cellLabelWidth =
-            this.spreadsheet.measureTextWidthRoughly(cellLabel);
+          const cellLabelWidth = this.spreadsheet.measureTextWidthRoughly(
+            cellLabel,
+            dataCellTextStyle,
+          );
 
           if (cellLabelWidth > maxDataLabelWidth) {
             maxDataLabel = cellLabel;